+++ src/components/Tabs/index.ts
@@ -1,1 +1,1 @@
-export { default as Tabs } from './Tabs';
+export { default as tabs } from "./tabs";
