+++ src/components/list/index.ts
@@ -1,1 +1,1 @@
-export { default as List } from './List';
+export { default as List } from "./list";
