+++ src/components/navigation-bar/index.ts
@@ -1,1 +1,1 @@
-export { default as NavigationBar } from './NavigationBar';
+export { default as NavigationBar } from './navigation-bar';
