+++ src/components/snack-bar/index.ts
@@ -1,0 +1,1 @@
+export { default as snackbar } from "./snack-bar";
