--- src/components/snackbar/index.ts
@@ -1,1 +1,0 @@
-export { default as SnackBar } from './Snackbar';
